refactor: trim BubbleRole and BubbleMetadata from models exports
